% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{b0fun}
\alias{b0fun}
\title{Calculate initial backward distribution}
\usage{
b0fun(g, r)
}
\arguments{
\item{g}{Intrinsic distribution}

\item{r}{discount rate}
}
\value{
backward distribution
}
\description{
Calculate an initial backward distribution by discounting and normalizing
an intrinsic distibution (or kernel)
}
