% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{genExp}
\alias{genExp}
\title{Generalized exponential}
\usage{
genExp(rho, kappa = NULL, degree = NULL)
}
\arguments{
\item{rho}{is the value to be exponentiated}

\item{kappa}{is 1/degree (don't specify both)}

\item{degree}{is the degree of the polynomial (extent of compounding)}
}
\description{
Generalized exponential
}
