% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{genLog}
\alias{genLog}
\title{Generalized logarithm}
\usage{
genLog(y, kappa = NULL, degree = NULL)
}
\arguments{
\item{y}{is the value to be logged}

\item{kappa}{is 1/degree (don't specify both)}

\item{degree}{is the degree of the polynomial (extent of compounding)}
}
\description{
Generalized logarithm
}
